fix(data_quality): resolve runtime issues in module integration
- Fix backend page routing: move pages to backend/data-quality/ prefix
  for correct URL generation (/backend/data-quality/checks, etc.)
- Fix relative import paths after page relocation (add extra ../ level)
- Fix API URL paths in UI components: /api/data-quality/ → /api/data_quality/
- Fix list filter bug: use != null guard for parseBooleanToken result
- Fix worker bootstrap: move target registry to shared package with
  registration in CLI bootstrap path (same pattern as search)
- Fix entity ID references in CRUD routes
- Add dataQualityTargetEntries to BootstrapData type
- Export dataQualityConfig from catalog and customers data-quality.ts

All endpoints verified: CRUD, scans, findings lifecycle, summary.
Scan worker runs correctly in queue process context.
Backend UI pages all return 200 with correct component rendering.
