+++ packages/core/src/modules/customers/data-quality.ts
@@ -1,6 +1,6 @@
 import type { DataQualityModuleConfig } from '@open-mercato/shared/modules/data-quality'
 
-const dataQualityConfig: DataQualityModuleConfig = {
+export const dataQualityConfig: DataQualityModuleConfig = {
   targets: [
     {
       entityId: 'customers:people',
