+++ packages/core/src/modules/data_quality/backend/data-quality/checks/page.tsx
@@ -12,41 +12,40 @@ import { raiseCrudError } from '@open-mercato/ui/backend/utils/serverErrors'
 import { Button } from '@open-mercato/ui/primitives/button'
 import { useT } from '@open-mercato/shared/lib/i18n/context'
 
-export { metadata } from './page.meta'
-
-type SuiteRow = {
+type CheckRow = {
   id: string
   code?: string | null
   name?: string | null
-  description?: string | null
+  targetEntityType?: string | null
+  severity?: string | null
   enabled?: boolean | null
 }
 
-type SuitesResponse = {
-  items?: SuiteRow[]
+type ChecksResponse = {
+  items?: CheckRow[]
   total?: number
   totalPages?: number
   page?: number
   pageSize?: number
 }
 
-export default function DataQualitySuitesPage() {
+export default function DataQualityChecksPage() {
   const t = useT()
   const [search, setSearch] = React.useState('')
   const [page, setPage] = React.useState(1)
   const [pageSize, setPageSize] = React.useState(25)
 
-  const { data, isLoading, isFetching, error, refetch } = useQuery<SuitesResponse>({
-    queryKey: ['data_quality_suites', search, page, pageSize],
+  const { data, isLoading, isFetching, error, refetch } = useQuery<ChecksResponse>({
+    queryKey: ['data_quality_checks', search, page, pageSize],
     queryFn: async () => {
       const params = new URLSearchParams({
         page: String(page),
         pageSize: String(pageSize),
       })
       if (search.trim()) params.set('search', search.trim())
-      const result = await apiCall<SuitesResponse>(`/api/data-quality/suites?${params.toString()}`)
+      const result = await apiCall<ChecksResponse>(`/api/data_quality/checks?${params.toString()}`)
       if (!result.ok) {
-        await raiseCrudError(result.response, t('data_quality.errors.suitesLoadFailed', 'Failed to load suites.'))
+        await raiseCrudError(result.response, t('data_quality.errors.checksLoadFailed', 'Failed to load checks.'))
       }
       return {
         items: Array.isArray(result.result?.items) ? result.result.items : [],
@@ -58,28 +57,32 @@ export default function DataQualitySuitesPage() {
     },
   })
 
-  const columns = React.useMemo<ColumnDef<SuiteRow>[]>(() => [
+  const columns = React.useMemo<ColumnDef<CheckRow>[]>(() => [
     {
       id: 'code',
       accessorKey: 'code',
-      header: t('data_quality.suites.columns.code', 'Code'),
+      header: t('data_quality.checks.columns.code', 'Code'),
     },
     {
       id: 'name',
       accessorKey: 'name',
-      header: t('data_quality.suites.columns.name', 'Name'),
+      header: t('data_quality.checks.columns.name', 'Name'),
+    },
+    {
+      id: 'targetEntityType',
+      accessorKey: 'targetEntityType',
+      header: t('data_quality.checks.columns.target', 'Target'),
+    },
+    {
+      id: 'severity',
+      accessorKey: 'severity',
+      header: t('data_quality.checks.columns.severity', 'Severity'),
     },
     {
       id: 'enabled',
       accessorKey: 'enabled',
-      header: t('data_quality.suites.columns.enabled', 'Enabled'),
+      header: t('data_quality.checks.columns.enabled', 'Enabled'),
       cell: ({ getValue }) => getValue() ? t('common.yes', 'Yes') : t('common.no', 'No'),
-    },
-    {
-      id: 'description',
-      accessorKey: 'description',
-      header: t('common.description', 'Description'),
-      cell: ({ getValue }) => getValue() ?? '—',
     },
   ], [t])
 
@@ -87,7 +90,7 @@ export default function DataQualitySuitesPage() {
     <Page>
       <PageBody>
         <DataTable
-          title={t('data_quality.suites.title', 'Data Quality Suites')}
+          title={t('data_quality.checks.title', 'Data Quality Checks')}
           data={data?.items ?? []}
           columns={columns}
           isLoading={isLoading}
@@ -97,7 +100,7 @@ export default function DataQualitySuitesPage() {
             setSearch(value)
             setPage(1)
           }}
-          searchPlaceholder={t('data_quality.suites.searchPlaceholder', 'Search suites')}
+          searchPlaceholder={t('data_quality.checks.searchPlaceholder', 'Search checks')}
           refreshButton={{
             label: t('common.refresh', 'Refresh'),
             onRefresh: () => { void refetch() },
@@ -105,13 +108,13 @@ export default function DataQualitySuitesPage() {
           }}
           actions={(
             <Button asChild>
-              <Link href="/backend/data-quality/suites/create">{t('data_quality.suites.create', 'Create Suite')}</Link>
+              <Link href="/backend/data-quality/checks/create">{t('data_quality.checks.create', 'Create Check')}</Link>
             </Button>
           )}
           rowActions={(row) => [{
-            id: 'edit-suite',
+            id: 'edit-check',
             label: t('common.edit', 'Edit'),
-            href: `/backend/data-quality/suites/${row.id}`,
+            href: `/backend/data-quality/checks/${row.id}`,
             icon: Pencil,
           }]}
           disableRowClick
@@ -128,8 +131,8 @@ export default function DataQualitySuitesPage() {
           }}
           emptyState={(
             <div className="py-8 text-center text-sm text-muted-foreground">
-              <p>{t('data_quality.suites.empty', 'No suites defined yet.')}</p>
-              <p>{t('data_quality.suites.emptyDescription', 'Create a suite to group checks together for batch scanning.')}</p>
+              <p>{t('data_quality.checks.empty', 'No checks defined yet.')}</p>
+              <p>{t('data_quality.checks.emptyDescription', 'Create your first data quality check to start monitoring your data.')}</p>
             </div>
           )}
         />
