+++ packages/core/src/modules/data_quality/backend/data-quality/suites/page.tsx
@@ -12,42 +12,39 @@ import { raiseCrudError } from '@open-mercato/ui/backend/utils/serverErrors'
 import { Button } from '@open-mercato/ui/primitives/button'
 import { useT } from '@open-mercato/shared/lib/i18n/context'
 
-export { metadata } from './page.meta'
-
-type CheckRow = {
+type SuiteRow = {
   id: string
   code?: string | null
   name?: string | null
-  targetEntityType?: string | null
-  severity?: string | null
+  description?: string | null
   enabled?: boolean | null
 }
 
-type ChecksResponse = {
-  items?: CheckRow[]
+type SuitesResponse = {
+  items?: SuiteRow[]
   total?: number
   totalPages?: number
   page?: number
   pageSize?: number
 }
 
-export default function DataQualityChecksPage() {
+export default function DataQualitySuitesPage() {
   const t = useT()
   const [search, setSearch] = React.useState('')
   const [page, setPage] = React.useState(1)
   const [pageSize, setPageSize] = React.useState(25)
 
-  const { data, isLoading, isFetching, error, refetch } = useQuery<ChecksResponse>({
-    queryKey: ['data_quality_checks', search, page, pageSize],
+  const { data, isLoading, isFetching, error, refetch } = useQuery<SuitesResponse>({
+    queryKey: ['data_quality_suites', search, page, pageSize],
     queryFn: async () => {
       const params = new URLSearchParams({
         page: String(page),
         pageSize: String(pageSize),
       })
       if (search.trim()) params.set('search', search.trim())
-      const result = await apiCall<ChecksResponse>(`/api/data-quality/checks?${params.toString()}`)
+      const result = await apiCall<SuitesResponse>(`/api/data_quality/suites?${params.toString()}`)
       if (!result.ok) {
-        await raiseCrudError(result.response, t('data_quality.errors.checksLoadFailed', 'Failed to load checks.'))
+        await raiseCrudError(result.response, t('data_quality.errors.suitesLoadFailed', 'Failed to load suites.'))
       }
       return {
         items: Array.isArray(result.result?.items) ? result.result.items : [],
@@ -59,32 +56,28 @@ export default function DataQualityChecksPage() {
     },
   })
 
-  const columns = React.useMemo<ColumnDef<CheckRow>[]>(() => [
+  const columns = React.useMemo<ColumnDef<SuiteRow>[]>(() => [
     {
       id: 'code',
       accessorKey: 'code',
-      header: t('data_quality.checks.columns.code', 'Code'),
+      header: t('data_quality.suites.columns.code', 'Code'),
     },
     {
       id: 'name',
       accessorKey: 'name',
-      header: t('data_quality.checks.columns.name', 'Name'),
-    },
-    {
-      id: 'targetEntityType',
-      accessorKey: 'targetEntityType',
-      header: t('data_quality.checks.columns.target', 'Target'),
-    },
-    {
-      id: 'severity',
-      accessorKey: 'severity',
-      header: t('data_quality.checks.columns.severity', 'Severity'),
+      header: t('data_quality.suites.columns.name', 'Name'),
     },
     {
       id: 'enabled',
       accessorKey: 'enabled',
-      header: t('data_quality.checks.columns.enabled', 'Enabled'),
+      header: t('data_quality.suites.columns.enabled', 'Enabled'),
       cell: ({ getValue }) => getValue() ? t('common.yes', 'Yes') : t('common.no', 'No'),
+    },
+    {
+      id: 'description',
+      accessorKey: 'description',
+      header: t('common.description', 'Description'),
+      cell: ({ getValue }) => getValue() ?? '—',
     },
   ], [t])
 
@@ -92,7 +85,7 @@ export default function DataQualityChecksPage() {
     <Page>
       <PageBody>
         <DataTable
-          title={t('data_quality.checks.title', 'Data Quality Checks')}
+          title={t('data_quality.suites.title', 'Data Quality Suites')}
           data={data?.items ?? []}
           columns={columns}
           isLoading={isLoading}
@@ -102,7 +95,7 @@ export default function DataQualityChecksPage() {
             setSearch(value)
             setPage(1)
           }}
-          searchPlaceholder={t('data_quality.checks.searchPlaceholder', 'Search checks')}
+          searchPlaceholder={t('data_quality.suites.searchPlaceholder', 'Search suites')}
           refreshButton={{
             label: t('common.refresh', 'Refresh'),
             onRefresh: () => { void refetch() },
@@ -110,13 +103,13 @@ export default function DataQualityChecksPage() {
           }}
           actions={(
             <Button asChild>
-              <Link href="/backend/data-quality/checks/create">{t('data_quality.checks.create', 'Create Check')}</Link>
+              <Link href="/backend/data-quality/suites/create">{t('data_quality.suites.create', 'Create Suite')}</Link>
             </Button>
           )}
           rowActions={(row) => [{
-            id: 'edit-check',
+            id: 'edit-suite',
             label: t('common.edit', 'Edit'),
-            href: `/backend/data-quality/checks/${row.id}`,
+            href: `/backend/data-quality/suites/${row.id}`,
             icon: Pencil,
           }]}
           disableRowClick
@@ -133,8 +126,8 @@ export default function DataQualityChecksPage() {
           }}
           emptyState={(
             <div className="py-8 text-center text-sm text-muted-foreground">
-              <p>{t('data_quality.checks.empty', 'No checks defined yet.')}</p>
-              <p>{t('data_quality.checks.emptyDescription', 'Create your first data quality check to start monitoring your data.')}</p>
+              <p>{t('data_quality.suites.empty', 'No suites defined yet.')}</p>
+              <p>{t('data_quality.suites.emptyDescription', 'Create a suite to group checks together for batch scanning.')}</p>
             </div>
           )}
         />
