+++ packages/core/src/modules/data_quality/di.ts
@@ -1,12 +1,10 @@
 import type { AppContainer } from '@open-mercato/shared/lib/di/container'
+import { loadTargetRegistry } from './lib/targetRegistry'
 
 export function register(container: AppContainer) {
   container.register({
     dataQualityTargetRegistry: {
-      resolve: () => {
-        const { loadTargetRegistry } = require('./lib/targetRegistry') as typeof import('./lib/targetRegistry')
-        return loadTargetRegistry()
-      },
+      resolve: () => loadTargetRegistry(),
     },
   })
 }
