+++ packages/shared/src/lib/bootstrap/dynamicLoader.ts
@@ -154,11 +154,13 @@ export async function loadBootstrapData(appRoot?: string): Promise<BootstrapData
     entitiesModule,
     diModule,
     searchModule,
+    dataQualityModule,
   ] = await Promise.all([
     compileAndImport(path.join(generatedDir, 'modules.cli.generated.ts')),
     compileAndImport(path.join(generatedDir, 'entities.generated.ts')),
     compileAndImport(path.join(generatedDir, 'di.generated.ts')),
     compileAndImport(path.join(generatedDir, 'search.generated.ts')).catch(() => ({ searchModuleConfigs: [] })),
+    compileAndImport(path.join(generatedDir, 'data-quality.generated.ts')).catch(() => ({ dataQualityTargetEntries: [] })),
   ])
 
   return {
@@ -168,6 +170,8 @@ export async function loadBootstrapData(appRoot?: string): Promise<BootstrapData
     entityIds: entityIdsModule.E as BootstrapData['entityIds'],
     // Search configs are needed by workers for indexing
     searchModuleConfigs: (searchModule.searchModuleConfigs ?? []) as BootstrapData['searchModuleConfigs'],
+    // Data quality target entries needed by scan workers
+    dataQualityTargetEntries: (dataQualityModule.dataQualityTargetEntries ?? []) as BootstrapData['dataQualityTargetEntries'],
     // Empty UI-related data - not needed for CLI
     dashboardWidgetEntries: [],
     injectionWidgetEntries: [],
